+++ pkgs/dart_mcp_server/lib/src/utils/names.dart
@@ -4,9 +4,10 @@
 
 /// A namespace for all the parameter names.
 extension ParameterNames on Never {
-  static const arguments = 'arguments';
+  static const applyFixes = 'applyFixes';
   static const apps = 'apps';
   static const appUri = 'appUri';
+  static const arguments = 'arguments';
   static const column = 'column';
   static const command = 'command';
   static const directory = 'directory';